Move app initialization from bin/skonfig to skonfig/__main__.py
This allows to run skonfig also as `python3 -m skonfig ...`.

Moreover, this is preparation for a potential future single file distribution of
skonfig not requiring pip or setuptools.
