+++ bin/skonfig
@@ -2,14 +2,9 @@
 
 if __name__ == "__main__":
     import os
-    skonfig_bindir_pardir = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
     import sys
+    skonfig_bindir_pardir = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
     if os.path.isfile(os.path.join(skonfig_bindir_pardir, "skonfig", "__init__.py")):
         sys.path.insert(0, skonfig_bindir_pardir)
-    import skonfig
-    try:
-        if skonfig.run():
-            sys.exit(0)
-        sys.exit(1)
-    except KeyboardInterrupt:
-        sys.exit(2)
+    import skonfig.__main__
+    skonfig.__main__.run()
