+++ skonfig/arguments.py
@@ -41,6 +41,7 @@ def get():
         "-j",
         dest="jobs",
         metavar="jobs",
+        type=int,
         help="maximum number of jobs (defaults to host CPU count, maximum: 4)"
     )
     parser.add_argument(
